TD-622: Various fixes to compare.js
* Removes the notion of a single mode, now we always use the continuous
mode in which the iframes are always visible
* Polyfills Promise API to support IE11 - haven't tested on earlier
versions
* Gecko is really slow at calculating the differences, which resulted
in the browser locking up completely and the "do you want to stop
this script" message was appearing. To counteract this, I have made
the loading of visual differences asynchronous (by using a timeout of
0ms), and given feedback to the user on the progress. It's interesting
just how more performant Trident and V8 are.
* Adds the possibility to add exclusions. This is provided with a JSON
block
* You can completely ignore a tag and it's children, or you can choose
    to simply ignore specified attribute changes to it. This is useful
    for things such as CSRF tokens and third-party social media content
* Ignores absolute URL style differences when the only thing which has
changed is the domain
* Doesn't report style changes when the element has the same signature
and the differences are identical - instead makes a note of the
number of elements changed
* Adds a timeout to the loading process. Will put this into a GUI
setting, but this prevents some silly Firefox bugs