Setting up Qunit test suite
* Using Node and Mocha would only go so far, as this is
a DOM heavy app. The best way to test this will be
using a fully-featured browser.
* At the minute I think that might be PhantomJS, and
there are ways of launching Qunit in it, and using
node app to output the messages to the console
* This will be my preferred approached
* At the minute I've just added to the build script the
ability to take a Qunit HTML template, and inject all the
test suites into it, creating an HTML file once it's done
* We'll then get PhantomJS to run the index.html file
and voila!